+++ R/reprex.R
@@ -20,7 +20,7 @@
 #' clipboard, and written to file. An HTML preview displays in RStudio's Viewer
 #' pane, if available, or in the default browser, otherwise. Leading `"> "`
 #' prompts, are stripped from the input code. Read more at
-#' <http://reprex.tidyverse.org/>.
+#' <https://reprex.tidyverse.org/>.
 #'
 #' reprex sets specific [knitr options](http://yihui.name/knitr/options/):
 #' * Chunk options default to `collapse = TRUE`, `comment = "#>"`,