+++ inst/templates/REPREX.R
@@ -12,9 +12,7 @@ knitr::opts_knit$set(upload.fun = {{{upload_fun}}})
 
 {{{std_file_stub}}}
 
-{{#advertise}}
-#' Created on `r Sys.Date()` by the [reprex package](https://reprex.tidyverse.org) (v`r utils::packageVersion("reprex")`).
-{{/advertise}}
+{{{ad}}}
 
 {{#si}}
 {{{si_start}}}