+++ NEWS.md
@@ -1,4 +1,4 @@
-# reprex 0.2.0.9000
+# reprex 0.2.1
 
 * The reprex ad is formatted as superscript for `venue = "gh"` and `venue = "so"`, i.e. it is more subtle (#201).